requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosmtpd==1.4.6
altgraph==0.17.4
amqp==5.3.1
asgiref==3.8.1
atpublic==5.1
attrs==25.3.0
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
cron-descriptor==1.4.5
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.1.7
django-allauth==65.5.0
django-celery-beat==2.7.0
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-filter==25.1
django-js-asset==3.1.2
django-slick-reporting==1.3.1
django-timezone-field==7.1
djangorestframework==3.15.2
frozenlist==1.5.0
geographiclib==2.0
geoip2==5.0.1
geopy==2.4.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
kombu==5.5.0
maxminddb==2.6.3
multidict==6.2.0
packaging==24.2
pefile==2023.2.7
phonenumbers==9.0.1
pillow==11.1.0
pip-review==1.3.0
polib==1.2.0
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2==2.9.10
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==77.0.3
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
twilio==9.5.1
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.18.3
